requirements.txt
click==7.1.2
requests==2.25.1
python-decouple==3.4
